requirements.txt
SpeechRecognition==3.14.3
pyttsx3==2.90
pyaudio
wikipedia
google.generativeai
dotenv
pyautogui